Add exports, move example code to file
