+++ zui/src/ui/stories/example.tsx
@@ -1,5 +1,6 @@
 import { Zui, zui as zuiImport } from '../../zui'
 import { defaultExtensions } from '../defaultextension'
+import { ZuiComponentMap } from '../types'
 const zui = zuiImport as Zui<typeof defaultExtensions>
 
 export const exampleSchema = zui
@@ -55,3 +56,43 @@ export const exampleSchema = zui
   })
   .displayAs('group', {})
   .title('User Information')
+
+export const componentMap: ZuiComponentMap<typeof defaultExtensions> = {
+  string: {
+    datetimeinput: () => null,
+    textbox: ({ params, onChange, errors, required, label, data }) => (
+      <div style={{ padding: '1rem' }}>
+        <span>{label}</span>
+        <input
+          placeholder={params.fitContentWidth ? 'fitContentWidth' : 'default'}
+          onChange={(e) => onChange(e.target.value)}
+          type={params.multiline ? 'textarea' : 'text'}
+        />
+        {required && <span>*</span>}
+        {errors && typeof data !== 'undefined' && <span style={{ color: 'red' }}>{errors}</span>}
+      </div>
+    ),
+    default: () => null,
+  },
+  array: {
+    default: () => null,
+    select: ({ children }) => <div>array: {children}</div>,
+  },
+  boolean: {
+    checkbox: () => null,
+    default: () => null,
+  },
+  number: {
+    default: () => null,
+    numberinput: () => null,
+    slider: () => null,
+  },
+  object: {
+    category: () => null,
+    default: () => null,
+    verticalLayout: () => null,
+    horizontalLayout: () => null,
+    group: ({ children }) => <div>{children}</div>,
+    categorization: () => null,
+  },
+}
